deselect structure if units are selected
commit is a cosmetic fix: previously,
when moving units around or otherwise
ordering them, the selection rectangle of
a structure kept pulsating white/gray.

this was confusing, since the players
mouseUnitsSelectedState does not inter-
act with that (structure) selection
and the player does not order the
structure in combination with units
selected.