leave RepairState with CTRL, SHIFT, E key if used after it's entered
these set of keys do not have a meaning while in repair state,
but trigger behavior in other states.  this change assumes the
player wants to leave to the previous state and have that re-
act to these keys.

It skips the need to press 'R' to leave RepairState in
order to use these keys in context of the previous state.